+++ changelog.md
@@ -6,11 +6,13 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 
 ## [Unreleased]
 
+## [Released]
+
+## [2.11.0] = 2021-01-29
+
 ### Added
 
 - Add user password change modal
-
-## [Released]
 
 ## [2.10.3] - 2021-01-28